+++ Pasteboard/Utilities/VisualConfiguration.swift
@@ -10,6 +10,6 @@ enum VisualConfiguration {
     static let rowActionSize: CGFloat = 28
     static let thumbnailSize = CGSize(width: 56, height: 56)
     static let searchFieldSpacing: CGFloat = 12
-    static let emptyStatePadding: CGFloat = 32
+    static let emptyStateTopPadding: CGFloat = 24
     static let quickAnimationDuration: TimeInterval = 0.16
 }